libsel4: Make FORCE_LONG_ENUM actually force a long instead of an int

+++ libsel4/include/sel4/macros.h
@@ -16,10 +16,10 @@
 /*
  * Some compilers attempt to pack enums into the smallest possible type.
  * For ABI compatability with the kernel, we need to ensure they remain
- * the same size as an 'int'.
+ * the same size as a 'long'.
  */
 #define SEL4_FORCE_LONG_ENUM(type) \
-        _enum_pad_ ## type = (1U << ((sizeof(int)*8) - 1)) - 1
+        _enum_pad_ ## type = (1ULL << ((sizeof(long)*8) - 1)) - 1
 
 #ifndef CONST
 #define CONST        __attribute__((__const__))
